Add X-Purism-FormFactor to desktop files
Without a properly set X-Purism-FormFactor value,
the Phosh (Wayland shell for GNOME) does not show the
the app, unless the "Show only adaptive apps" radio button
is unticked. Since the GUI of osmscout-server is fully adaptive
and ready to be used, we should declare the form factor to be
mobile ready.