Adjust path colour in mc.json to contrast with background
In the mc / "Midnight Commander" (night) style footpaths were rendered
using the same colour as the generic background. As such they were only
visible when they crossed areas with a different coloured background,
such as woodland, AONBs, etc. This changes them to use a colour which is
visible on all backgrounds, but still in keeping with the colour scheme.